requirements.txt
# 核心依赖
onnxruntime==1.16.3
sentence-transformers==2.3.1
chromadb==0.4.22
llama-index==0.9.26

# 文档处理
unidecode==1.3.7
pymupdf==1.23.7
python-docx==1.0.1
regex==2023.10.3
openpyxl==3.1.2
python-pptx==1.0.1

# 数据处理
numpy==1.26.2
pandas==2.1.4
tqdm==4.66.1

# API服务
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6

# Web界面
streamlit==1.30.0

# 系统工具
redis==5.0.1
watchdog==3.0.0
